Add comments for performance refactor
Currently my solution would have a serious performance issue since
it would literally rotate an array 1,000,000 if asked to. However,
that is totally unnecessary if K is a multiple of N. So I would
need to add a conditional to check for the remainder of N when
divided by K, then only perform that many actual rotations.